chore: Update project.properties and VideoPanel.java
This commit updates the project.properties file to include references to the JavaFX library files. It also modifies the VideoPanel.java file by adding a new method to check if the video is running and setting a listener for when the video is ready. Additionally, it fixes a bug in the salirDelVideo() method to dispose the mediaPlayer only if it is not null.